requirements.txt
# General
# ------------------------------------------------------------------------------
requests==2.31.0
python-slugify==8.0.1
Pillow==10.0.0
argon2-cffi==21.3.0
phonenumbers==8.13.17
ua-parser==0.18.0
user-agents==2.2.0

# Django & Django Helpers
# ------------------------------------------------------------------------------
django==4.2.3
django-environ==0.10.0
django-model-utils==4.3.1
django-redis==5.3.0
django-cors-headers==4.2.0
django-storages[azure]==1.13.2
django-anymail[sendgrid]==10.0
django-phonenumber-field[phonenumbers]==7.1.0
django-user-agents==0.4.0
django-extensions==3.2.3
django-choices==2.0.0

# Production
# ------------------------------------------------------------------------------
gunicorn==21.2.0
uvicorn[standard]==0.23.1

# Celery
# ------------------------------------------------------------------------------
celery==5.3.1
django-celery-beat==2.5.0

# Cache
# ------------------------------------------------------------------------------
redis==4.6.0
hiredis==2.2.3

# Rest Framework
# ------------------------------------------------------------------------------
djangorestframework==3.14.0
django-filter==23.2

# Database
# ------------------------------------------------------------------------------
psycopg2==2.9.6

# Channels / WebSocket
# ------------------------------------------------------------------------------
channels[daphne]==4.0.0
channels_redis==4.1.0


# KeyCloak Token Handling
# ------------------------------------------------------------------------------
pyjwt==2.8.0
